+++ jar/src/main/resources/application-context.xml
@@ -20,8 +20,8 @@
     <bean id="dataSource" class="org.springframework.jdbc.datasource.DriverManagerDataSource">
         <property name="driverClassName" value="${ds.database-driver}"/>
         <property name="url" value="${ds.url}"/>
-        <property name="username" value="${ds.username}"/>
-        <property name="password" value="${ds.password}"/>
+        <property name="username" value="student"/>
+        <property name="password" value="student"/>
     </bean>
 
     <bean id = "jdbcTemplate" class = "org.springframework.jdbc.core.JdbcTemplate">
